+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add `-l` flag for optional SAO logo
 - Add `-u` flag to update known threads but not scrape
 - Add `-p` flag to force using 4plebs for thread JSONs and attachments source
+- Add `-o` flag for oneoff config-independent thread scraping
 
 ## v3.0.5
 - Remove saosuite submodule since it is taking a different direction (private repo for my resources)